pyproject.toml
[project]
name = "prispevky"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jan Mrkos", email = "mrkosja1@fel.cvut.cz" }
]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.184.0",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "qrcode>=8.2",
    "unidecode>=1.4.0",
]

[project.scripts]
prispevky = "prispevky:main"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
